fix: skip unchanged image translations
